requirements.txt
dbt-postgres==1.8.2
sqlfluff==3.3.1
# sqlfluff-templater-dbt==3.3.1
elementary-data==0.16.2
python-dotenv==1.0.1

